requirements.txt
Django==1.4.2
South==0.7.6
argparse==1.2.1
dj-database-url==0.2.1
django-extensions==0.9
django-registration==0.8
django-taggit==0.9.3
wsgiref==0.1.2
git+https://github.com/toastdriven/django-haystack.git@master#egg=django-haystack
git+https://github.com/tgdn/django-bootstrap-pagination.git#egg=django-bootstrap-pagination
requests
pyelasticsearch
PyYAML==3.10
django-debug-toolbar==0.9.4
django-nose==1.1
yolk
django-crispy-forms==1.2.0
unicode-slugify
ipython
PIL
django-storages==1.1.5
boto==2.6.0
Unipath==0.2.1
gunicorn==0.15.0
psycopg2==2.4.5
django-social-auth==0.7.9
django-templated-email==0.4.7
Fabric==1.5.1
